+++ maser/utils/cdf/converter/xlsx/vattrs.py
@@ -22,19 +22,6 @@ __all__ = ["add_vattr",
 
 # ________________ HEADER _________________________
 
-# Mandatory
-__version__ = ""
-__author__ = "Xavier Bonnin"
-__date__ = "2016-09-30"
-
-# Optional
-__license__ = ""
-__credit__ = [""]
-__maintainer__ = ""
-__email__ = ""
-__project__ = "MASER"
-__institute__ = "LESIA, Observatoire de Paris, LESIA"
-__changes__ = ""
 
 
 # ________________ Global Variables _____________
@@ -73,7 +60,7 @@ def add_var(xlsx, varname, entries,
     wb = lwb(xlsx)
 
     # Get sheet for variable attributes
-    ws = wb.get_sheet_by_name(VATT_SHEET_NAME)
+    ws = wb[VATT_SHEET_NAME]
 
     # Check if the variable exists
     rows = get_row(ws, varname, column="A")
@@ -111,7 +98,7 @@ def add_vattr(xlsx, attname, cdftype, entry,
     wb = lwb(xlsx)
 
     # Get sheet for variable attributes
-    ws = wb.get_sheet_by_name(VATT_SHEET_NAME)
+    ws = wb[VATT_SHEET_NAME]
 
     # Check if the variable attribute exists
     rows = get_row(ws, attname, column="B")
@@ -163,7 +150,7 @@ def set_vattr_entries(xlsx, attname, new_entry,
     wb = lwb(xlsx)
 
     # Get sheet for variable attributes
-    ws = wb.get_sheet_by_name(VATT_SHEET_NAME)
+    ws = wb[VATT_SHEET_NAME]
 
     # Get indices of row for the vattr
     rows = get_row(ws, attname, column="B")
@@ -198,7 +185,7 @@ def set_vattr_dtype(xlsx, attname, new_dtype,
     wb = lwb(xlsx)
 
     # Get sheet for variable attributes
-    ws = wb.get_sheet_by_name(VATT_SHEET_NAME)
+    ws = wb[VATT_SHEET_NAME]
 
     # Get indices of row for the vattr
     rows = get_row(ws, attname, column="B")
@@ -231,7 +218,7 @@ def rename_vattr(xlsx, old_attname, new_attname,
     wb = lwb(xlsx)
 
     # Get sheet for variable attributes
-    ws = wb.get_sheet_by_name(VATT_SHEET_NAME)
+    ws = wb[VATT_SHEET_NAME]
 
     # Get indices of row for the vattr
     rows = get_row(ws, old_attname, column="B")
@@ -264,7 +251,7 @@ def rm_vattr(xlsx, attname,
     wb = lwb(xlsx)
 
     # Get sheet for variable attributes
-    ws = wb.get_sheet_by_name(VATT_SHEET_NAME)
+    ws = wb[VATT_SHEET_NAME]
 
     # Get indices of row for the vattr
     rows = get_row(ws, attname, column="B")
@@ -304,7 +291,7 @@ def rm_var(xlsx, varname,
     wb = lwb(xlsx)
 
     # Get sheet for variable attributes
-    ws = wb.get_sheet_by_name(VATT_SHEET_NAME)
+    ws = wb[VATT_SHEET_NAME]
 
     # Get indices of row for the vattr
     rows = get_row(ws, varname, column="A")
@@ -335,7 +322,7 @@ def set_vattr_var(xlsx, old_varname, new_varname,
     wb = lwb(xlsx)
 
     # Get sheet for variable attributes
-    ws = wb.get_sheet_by_name(VATT_SHEET_NAME)
+    ws = wb[VATT_SHEET_NAME]
 
     # Check if the variable attribute exists
     rows = get_row(ws, old_varname)